+++ sdk/spring/azure-spring-boot-test-aad-b2c/src/test/java/com/azure/test/aad/b2c/selenium/AADB2CSeleniumITHelper.java
@@ -4,8 +4,8 @@ import static com.azure.spring.test.EnvironmentVariable.AAD_B2C_CLIENT_ID;
 import static com.azure.spring.test.EnvironmentVariable.AAD_B2C_CLIENT_SECRET;
 import static com.azure.spring.test.EnvironmentVariable.AAD_B2C_PROFILE_EDIT;
 import static com.azure.spring.test.EnvironmentVariable.AAD_B2C_REPLY_URL;
-import static com.azure.spring.test.EnvironmentVariable.AAD_B2C_SIGN_UP_OR_SIGN_IN;
-import static com.azure.spring.test.EnvironmentVariable.AAD_B2C_TENANT;
+import static com.azure.spring.test.EnvironmentVariable.AAD_B2C_SIGN_IN_USER_FLOW;
+import static com.azure.spring.test.EnvironmentVariable.AAD_B2C_BASE_URI;
 import static com.azure.spring.test.EnvironmentVariable.AAD_B2C_USER_EMAIL;
 import static com.azure.spring.test.EnvironmentVariable.AAD_B2C_USER_PASSWORD;
 import static org.openqa.selenium.support.ui.ExpectedConditions.presenceOfElementLocated;
@@ -24,14 +24,14 @@ public class AADB2CSeleniumITHelper extends SeleniumITHelper {
 
     public static Map<String, String> createDefaultProperteis() {
         Map<String, String> defaultProperteis = new HashMap<>();
-        defaultProperteis.put("azure.activedirectory.b2c.tenant", AAD_B2C_TENANT);
+        defaultProperteis.put("azure.activedirectory.b2c.base-uri", AAD_B2C_BASE_URI);
         defaultProperteis.put("azure.activedirectory.b2c.client-id", AAD_B2C_CLIENT_ID);
         defaultProperteis.put("azure.activedirectory.b2c.client-secret", AAD_B2C_CLIENT_SECRET);
         defaultProperteis.put("azure.activedirectory.b2c.reply-url", AAD_B2C_REPLY_URL);
         defaultProperteis
-            .put("azure.activedirectory.b2c.user-flows.profile-edit", AAD_B2C_PROFILE_EDIT);
+            .put("azure.activedirectory.b2c.sign-in-user-flow", AAD_B2C_SIGN_IN_USER_FLOW);
         defaultProperteis
-            .put("azure.activedirectory.b2c.user-flows.sign-up-or-sign-in", AAD_B2C_SIGN_UP_OR_SIGN_IN);
+            .put("azure.activedirectory.b2c.user-flows", AAD_B2C_PROFILE_EDIT);
         return defaultProperteis;
     }
 
@@ -61,7 +61,7 @@ public class AADB2CSeleniumITHelper extends SeleniumITHelper {
         wait.until(ExpectedConditions.elementToBeClickable(By.cssSelector("button[type='submit']"))).submit();
         manualRedirection();
         wait.until(ExpectedConditions.elementToBeClickable(By.cssSelector(
-            "a[href='/oauth2/authorization/" + AAD_B2C_SIGN_UP_OR_SIGN_IN + "']"))).click();
+            "a[href='/oauth2/authorization/" + AAD_B2C_SIGN_IN_USER_FLOW + "']"))).click();
     }
 
     private void manualRedirection() {