+++ sdk/spring/azure-spring-boot/src/main/java/com/azure/spring/autoconfigure/b2c/AADB2CAuthorizationRequestResolver.java
@@ -2,6 +2,8 @@
 // Licensed under the MIT License.
 package com.azure.spring.autoconfigure.b2c;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.lang.NonNull;
 import org.springframework.lang.Nullable;
 import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
@@ -26,6 +28,8 @@ import java.util.Optional;
  */
 public class AADB2CAuthorizationRequestResolver implements OAuth2AuthorizationRequestResolver {
 
+    private static final Logger LOGGER = LoggerFactory.getLogger(AADB2CAuthorizationRequestResolver.class);
+
     private static final String REQUEST_BASE_URI =
             OAuth2AuthorizationRequestRedirectFilter.DEFAULT_AUTHORIZATION_REQUEST_BASE_URI;
 
@@ -41,14 +45,11 @@ public class AADB2CAuthorizationRequestResolver implements OAuth2AuthorizationRe
 
     private final OAuth2AuthorizationRequestResolver defaultResolver;
 
-    private final String passwordResetUserFlow;
-
     private final AADB2CProperties properties;
 
     public AADB2CAuthorizationRequestResolver(@NonNull ClientRegistrationRepository repository,
                                               @NonNull AADB2CProperties properties) {
         this.properties = properties;
-        this.passwordResetUserFlow = this.properties.getUserFlows().getPasswordReset();
         this.defaultResolver = new DefaultOAuth2AuthorizationRequestResolver(repository, REQUEST_BASE_URI);
     }
 
@@ -59,13 +60,17 @@ public class AADB2CAuthorizationRequestResolver implements OAuth2AuthorizationRe
 
     @Override
     public OAuth2AuthorizationRequest resolve(@NonNull HttpServletRequest request, String registrationId) {
-        if (StringUtils.hasText(passwordResetUserFlow) && isForgotPasswordAuthorizationRequest(request)) {
-            final OAuth2AuthorizationRequest authRequest = defaultResolver.resolve(request, passwordResetUserFlow);
-            return getB2CAuthorizationRequest(authRequest, passwordResetUserFlow);
-        }
+        if (StringUtils.hasText(registrationId)) {
+            if (REQUEST_MATCHER.matches(request)) {
+                return getB2CAuthorizationRequest(defaultResolver.resolve(request), registrationId);
+            }
 
-        if (StringUtils.hasText(registrationId) && REQUEST_MATCHER.matches(request)) {
-            return getB2CAuthorizationRequest(defaultResolver.resolve(request), registrationId);
+            if (isForgotPasswordAuthorizationRequest(request)) {
+                final OAuth2AuthorizationRequest authRequest = defaultResolver.resolve(request, registrationId);
+                return getB2CAuthorizationRequest(authRequest, registrationId);
+            }
+
+            LOGGER.warn("Unrecognized registration id: {}", registrationId);
         }
 
         // Return null may not be the good practice, but we need to align with oauth2.client.web