+++ sdk/spring/azure-spring-boot/src/main/java/com/azure/spring/autoconfigure/b2c/AADB2CProperties.java
@@ -6,12 +6,12 @@ import org.hibernate.validator.constraints.URL;
 import org.springframework.beans.factory.InitializingBean;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.DeprecatedConfigurationProperty;
-import org.springframework.security.oauth2.client.authentication.OAuth2AuthorizationCodeAuthenticationProvider;
-import org.springframework.security.oauth2.client.oidc.authentication.OidcAuthorizationCodeAuthenticationProvider;
 import org.springframework.util.StringUtils;
 import org.springframework.validation.annotation.Validated;
 
 import javax.validation.constraints.NotBlank;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -23,21 +23,13 @@ import java.util.regex.Pattern;
 @ConfigurationProperties(prefix = AADB2CProperties.PREFIX)
 public class AADB2CProperties implements InitializingBean {
 
-    private static final String USER_FLOWS = "user-flows";
+    public static final String USER_FLOWS = "user-flows";
 
     /**
-     * We do not use ${@link String#format(String, Object...)} as it's not real constant, which cannot be referenced in
-     * annotation.
+     * We do not use ${@link String#format(String, Object...)}
+     * as it's not real constant, which cannot be referenced in annotation.
      */
-    public static final String USER_FLOW_PASSWORD_RESET = USER_FLOWS + ".password-reset";
-
-    public static final String USER_FLOW_PROFILE_EDIT = USER_FLOWS + ".profile-edit";
-
-    public static final String USER_FLOW_SIGN_UP_OR_SIGN_IN = USER_FLOWS + ".sign-up-or-sign-in";
-
-    public static final String USER_FLOW_SIGN_UP = USER_FLOWS + ".sign-up";
-
-    public static final String USER_FLOW_SIGN_IN = USER_FLOWS + ".sign-in";
+    public static final String SIGN_IN_USER_FLOW = "sign-in-user-flow";
 
     public static final String DEFAULT_LOGOUT_SUCCESS_URL = "http://localhost:8080/login";
 
@@ -51,12 +43,6 @@ public class AADB2CProperties implements InitializingBean {
      */
     @Deprecated
     private String tenant;
-
-    /**
-     * Use OIDC ${@link OidcAuthorizationCodeAuthenticationProvider} by default. If set to false, will use Oauth2
-     * ${@link OAuth2AuthorizationCodeAuthenticationProvider}.
-     */
-    private Boolean oidcEnabled = true;
 
     /**
      * The application ID that registered under b2c tenant.
@@ -81,11 +67,6 @@ public class AADB2CProperties implements InitializingBean {
     private String userNameAttributeName;
 
     /**
-     * The all user flows which is created under b2c tenant.
-     */
-    private UserFlows userFlows = new UserFlows();
-
-    /**
      * Telemetry data will be collected if true, or disable data collection.
      */
     private boolean allowTelemetry = true;
@@ -98,87 +79,26 @@ public class AADB2CProperties implements InitializingBean {
     @URL(message = "baseUri should be valid URL")
     private String baseUri;
 
+    /**
+     * The all user flows which is created under b2c tenant.
+     */
+    private List<String> userFlows = new ArrayList<>();
+
+    /**
+     * Specify the primary sign in flow name
+     */
+    @NotBlank(message = "The primary sign in flow name should not be blank.")
+    private String signInUserFlow;
+
     @Override
     public void afterPropertiesSet() {
         if (StringUtils.isEmpty(tenant) && StringUtils.isEmpty(baseUri)) {
             throw new AADB2CConfigurationException("'tenant' and 'baseUri' at least configure one item.");
         }
-    }
 
-    /**
-     * UserFlows
-     */
-    @Validated
-    public static class UserFlows {
-
-        protected UserFlows() {
-
-        }
-
-        /**
-         * The sign-up-or-sign-in user flow which is created under b2c tenant.
-         */
-        @NotBlank(message = "sign-up-or-in value should not be blank")
-        private String signUpOrSignIn;
-
-        /**
-         * The profile-edit user flow which is created under b2c tenant.
-         */
-        private String profileEdit;
-
-        /**
-         * The password-reset user flow which is created under b2c tenant.
-         */
-        private String passwordReset;
-
-        /**
-         * The sign-up user flow which is created under b2c tenant.
-         */
-        private String signUp;
-
-        /**
-         * The sign-in user flow which is created under b2c tenant.
-         */
-        private String signIn;
-
-        public String getSignUp() {
-            return signUp;
-        }
-
-        public void setSignUp(String signUp) {
-            this.signUp = signUp;
-        }
-
-        public String getSignIn() {
-            return signIn;
-        }
-
-        public void setSignIn(String signIn) {
-            this.signIn = signIn;
-        }
-
-        public String getSignUpOrSignIn() {
-            return signUpOrSignIn;
-        }
-
-        public void setSignUpOrSignIn(String signUpOrSignIn) {
-            this.signUpOrSignIn = signUpOrSignIn;
-        }
-
-        public String getProfileEdit() {
-            return profileEdit;
-        }
-
-        public void setProfileEdit(String profileEdit) {
-            this.profileEdit = profileEdit;
-        }
-
-        public String getPasswordReset() {
-            return passwordReset;
-        }
-
-        public void setPasswordReset(String passwordReset) {
-            this.passwordReset = passwordReset;
+        if (!userFlows.isEmpty() && userFlows.contains(signInUserFlow)) {
+            throw new AADB2CConfigurationException("Sign in user flow '" + signInUserFlow
+                + "' does not need to be configured repeatedly.");
         }
     }
 
@@ -218,12 +138,20 @@ public class AADB2CProperties implements InitializingBean {
         return tenant;
     }
 
-    public Boolean getOidcEnabled() {
-        return oidcEnabled;
+    public List<String> getUserFlows() {
+        return userFlows;
     }
 
-    public void setOidcEnabled(Boolean oidcEnabled) {
-        this.oidcEnabled = oidcEnabled;
+    public void setUserFlows(List<String> userFlows) {
+        this.userFlows = userFlows;
+    }
+
+    public String getSignInUserFlow() {
+        return signInUserFlow;
+    }
+
+    public void setSignInUserFlow(String signInUserFlow) {
+        this.signInUserFlow = signInUserFlow;
     }
 
     public String getClientId() {
@@ -256,14 +184,6 @@ public class AADB2CProperties implements InitializingBean {
 
     public void setAuthenticateAdditionalParameters(Map<String, Object> authenticateAdditionalParameters) {
         this.authenticateAdditionalParameters = authenticateAdditionalParameters;
-    }
-
-    public UserFlows getUserFlows() {
-        return userFlows;
-    }
-
-    public void setUserFlows(UserFlows userFlows) {
-        this.userFlows = userFlows;
     }
 
     public boolean isAllowTelemetry() {